% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_db_handling.R
\name{append_row_db}
\alias{append_row_db}
\title{Add a labeled matrix containing the handwritten box data to the database}
\usage{
append_row_db(mat, label)
}
\arguments{
\item{mat}{numeric matrix containing the handwritten box data}

\item{label}{label either x for No or y for Yes (character)}
}
\value{
does not return something
}
\description{
Add a labeled matrix containing the handwritten box data to the database
}
\author{
Emanuel Sommer
}
