% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{average_no_plot}
\alias{average_no_plot}
\title{The average no box plot}
\format{
An object of class \code{gg} (inherits from \code{ggplot}) of length 9.
}
\usage{
average_no_plot
}
\description{
A ggplot2 object that visualizes the average no box.
}
\keyword{datasets}
