% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{build_matrix}
\alias{build_matrix}
\title{Build a 28x28 matrix from the two coordinate vectors x and y corresponding to the two axis of the Yes and No boxes.}
\usage{
build_matrix(x, y)
}
\arguments{
\item{x}{numeric vector with domain [1,28] / corresponding to the x axis of the box}

\item{y}{numeric vector with domain [1,28] / corresponding to the y axis of the box}
}
\value{
a 28x28 matrix of 0 and 1
}
\description{
First round the vectors and then output binary matrix indicating which points are drawn (the pixels)
}
\examples{
build_matrix(c(1,1,3.3,5,27.9),c(1,1,7.000012,5.555,28))
}
\author{
Emanuel Sommer
}
