% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_db_handling.R
\name{get_current_db}
\alias{get_current_db}
\title{Return current state of database as dataframe}
\usage{
get_current_db()
}
\value{
dataframe with the current state of the database
}
\description{
Return current state of database as dataframe
}
\author{
Emanuel Sommer
}
