% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_db_handling.R
\name{get_current_labels}
\alias{get_current_labels}
\title{Return current labels of database as dataframe}
\usage{
get_current_labels()
}
\value{
chr strin with the current labels of the database
}
\description{
Return current labels of database as dataframe
}
\author{
Emanuel Sommer
}
