% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_eda_db.R
\name{plot_average_box}
\alias{plot_average_box}
\title{Plot the average handdrawn boxes}
\usage{
plot_average_box(df, sel_label = c("x", "y"))
}
\arguments{
\item{df}{data from the database (tibble/dataframe with column 'label')}

\item{sel_label}{character label either x or y}
}
\value{
ggplot2 object
}
\description{
Plot the average handdrawn boxes
}
\author{
Emanuel Sommmer
}
