% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_eda_db.R
\name{plot_label_dist}
\alias{plot_label_dist}
\title{Plot the proportions of labels in the database as a barplot}
\usage{
plot_label_dist(data)
}
\arguments{
\item{data}{data from the database (tibble/dataframe with column 'label')}
}
\value{
ggplot2 object
}
\description{
Plot the proportions of labels in the database as a barplot
}
\author{
Emanuel Sommer
}
